feat: add castling movement
Refactored game.move_piece to include a unique
piece move, castling, which has different
logic to normal move.

Placed the coordinates of rook from which a king
is supposed to castle on Piece superclass, to be
used by King and Piece itself.

The Piece then uses #castle method for Game to
switch Rook and King pieces together.