feat: add checks & force player to use legal moves
One of the core component for this feature is
the undo method, which is done to validate
if the move does not leaves the king in check.
I will refactor this and create a solution that
makes a copy of the board instead of undoing,
which can complicate things further when I start
to consider other features like castling,
promotion, etc.