refactor: change process of checks and interaction
Instead of undoing the movement, I made a deep
clone of the board, use the selected move to that
clone, validate if the player is still in check.
If checked, then the original board stays the
same. If not checked, then we use the same move
in the original. Either way, the clone just acts
as a check simulator.

Refactored determining move action, either castle
or normal, and used a chessboard parameter to
determine whether use the original or clone.
Afterwards, refactored #move in main.rb due to
huge if-else block. More refactoring needed.