refactor: logic for Pawn#en_passant functionality
Instead of refreshing the en passant state in Game
class, I rebuilt the en passant to signal an
opposing adjacent pawn, provided that it checks
the logic. That signaled pawn will enable its
state to true, indicating that particular pawn can
en passant an opposing pawn.

On the other hand, the signaller is also a state
in itself. It will disable (falsify) the signal
and signaller automatically for every move.

I've thought of falsifying these immediately
before the moves are generated, because I am
skeptical that these states might still be true
and keep the en passant square, which should
disappear after a move has made.